Refactor migration test files to reuse common code
- Extract common utilities to migrationTestUtils.ts
- Refactor DashboardMigratorSingleVersion.test.ts to use shared utilities
- Refactor DashboardMigratorToBackend.test.ts to use shared utilities
- Reduce code duplication and improve maintainability
- All existing tests continue to pass
